// frontend/src/api/generated/models/UserDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserDto = {
    createdAt?: string;
    createdBy?: string | null;
    updatedAt?: string | null;
    updatedBy?: string | null;
    isDeleted?: boolean;
    deletedAt?: string | null;
    deletedBy?: string | null;
    id?: number;
    userName?: string | null;
    role?: string | null;
    isActive?: boolean;
};